refactor: moved some cli code from __main__.py to __init__.py
code related to formatting is moved to __init__.py. subparsers and
arguments are still in __main__.py
